refactor(web-ui): slice board and task detail surfaces
