+++ apps/web-ui/src/components/project-board.tsx
@@ -25,7 +25,7 @@ import {
 import { CSS } from '@dnd-kit/utilities';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useCallback, useEffect, useMemo, useRef, useState, type PointerEvent as ReactPointerEvent, type ReactNode } from 'react';
-import { Calendar, Check, CheckCircle2, ChevronDown, ChevronRight, Circle, Diamond, ExternalLink, Folder, Plus, Stamp, Trash2, User } from 'lucide-react';
+import { Calendar, Check, CheckCircle2, ChevronDown, ChevronRight, Circle, ExternalLink, Folder, Plus, Trash2, User } from 'lucide-react';
 import { api } from '@/lib/api';
 import { queryKeys } from '@/lib/query-keys';
 import type {
@@ -54,177 +54,23 @@ import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover
 import { Table, TableBody, TableCell, TableHeader, TableRow } from '@/components/ui/table';
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
 import { cn } from '@/lib/utils';
+import { useProjectBoardState } from '@/components/project-board/project-board-state';
+import {
+  customFieldColumnKey,
+  findTaskCustomFieldValue,
+  initials,
+  optimisticCustomFieldValues,
+  removeTaskFromGroups,
+  renderTaskTypeCompletionIcon,
+  resolveAssigneeLabel,
+  taskMatchesCustomFieldFilter,
+  upsertTaskInSection,
+} from '@/components/project-board/project-board-utils';
 import TaskDetailDrawer from '@/components/task-detail-drawer';
 import { useI18n } from '@/lib/i18n';
 import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/dialog';
 
-function sortByPosition(tasks: Task[]) {
-  return [...tasks].sort((a, b) => a.position - b.position);
-}
-
 const EMPTY_CUSTOM_FIELDS: CustomFieldDefinition[] = [];
-
-function renderTaskTypeCompletionIcon(task: Task, isDone: boolean) {
-  if (task.type === 'MILESTONE') {
-    return (
-      <Diamond
-        className={cn('h-5 w-5', isDone ? 'fill-current text-emerald-600' : 'text-muted-foreground')}
-      />
-    );
-  }
-  if (task.type === 'APPROVAL') {
-    return <Stamp className={cn('h-5 w-5', isDone ? 'text-emerald-600' : 'text-muted-foreground')} />;
-  }
-  return isDone ? <CheckCircle2 className="h-5 w-5" /> : <Circle className="h-5 w-5" />;
-}
-
-function removeTaskFromGroups(groups: SectionTaskGroup[], taskId: string) {
-  return groups.map((group) => ({
-    ...group,
-    tasks: group.tasks.filter((task) => task.id !== taskId),
-  }));
-}
-
-function upsertTaskInSection(groups: SectionTaskGroup[], sectionId: string, task: Task) {
-  return groups.map((group) => {
-    if (group.section.id !== sectionId) return group;
-    const nextTasks = sortByPosition([task, ...group.tasks.filter((item) => item.id !== task.id)]);
-    return { ...group, tasks: nextTasks };
-  });
-}
-
-function moveTaskPreview(
-  groups: SectionTaskGroup[],
-  taskId: string,
-  toSectionId: string,
-  targetTaskId: string | null,
-) {
-  const cloned = groups.map((group) => ({ ...group, tasks: [...group.tasks] }));
-  let movingTask: Task | null = null;
-
-  for (const group of cloned) {
-    const index = group.tasks.findIndex((task) => task.id === taskId);
-    if (index >= 0) {
-      movingTask = group.tasks[index] ?? null;
-      group.tasks.splice(index, 1);
-      break;
-    }
-  }
-
-  if (!movingTask) return groups;
-
-  const destination = cloned.find((group) => group.section.id === toSectionId);
-  if (!destination) return groups;
-
-  const insertAt = targetTaskId
-    ? Math.max(destination.tasks.findIndex((task) => task.id === targetTaskId), 0)
-    : destination.tasks.length;
-
-  destination.tasks.splice(insertAt, 0, { ...movingTask, sectionId: toSectionId });
-  return cloned;
-}
-
-function resolveAssigneeLabel(task: Task, members: ProjectMember[]) {
-  if (!task.assigneeUserId) return 'unassigned';
-  const member = members.find((item) => item.userId === task.assigneeUserId);
-  if (!member) return task.assigneeUserId;
-  return member.user.displayName || member.user.email || member.userId;
-}
-
-function initials(label: string) {
-  const pieces = label.trim().split(/\s+/).slice(0, 2);
-  return pieces.map((piece) => piece.charAt(0).toUpperCase()).join('') || 'U';
-}
-
-function customFieldColumnKey(fieldId: string) {
-  return `cf:${fieldId}`;
-}
-
-function findTaskCustomFieldValue(task: Task, fieldId: string): TaskCustomFieldValue | null {
-  return task.customFieldValues?.find((value) => value.fieldId === fieldId) ?? null;
-}
-
-function taskMatchesCustomFieldFilter(task: Task, filter: CustomFieldFilter): boolean {
-  const value = findTaskCustomFieldValue(task, filter.fieldId);
-  if (!value) return false;
-
-  if (filter.type === 'SELECT') {
-    if (!filter.optionIds?.length) return true;
-    return Boolean(value.optionId && filter.optionIds.includes(value.optionId));
-  }
-
-  if (filter.type === 'BOOLEAN') {
-    if (typeof filter.booleanValue !== 'boolean') return true;
-    return value.valueBoolean === filter.booleanValue;
-  }
-
-  if (filter.type === 'NUMBER') {
-    if (typeof value.valueNumber !== 'number') return false;
-    if (typeof filter.numberMin === 'number' && value.valueNumber < filter.numberMin) return false;
-    if (typeof filter.numberMax === 'number' && value.valueNumber > filter.numberMax) return false;
-    return true;
-  }
-
-  const valueDate = value.valueDate ? String(value.valueDate).slice(0, 10) : '';
-  if (!valueDate) return false;
-  if (filter.dateFrom && valueDate < filter.dateFrom) return false;
-  if (filter.dateTo && valueDate > filter.dateTo) return false;
-  return true;
-}
-
-function optimisticCustomFieldValues(
-  task: Task,
-  field: CustomFieldDefinition,
-  rawValue: unknown,
-): TaskCustomFieldValue[] {
-  const current = [...(task.customFieldValues ?? [])];
-  const index = current.findIndex((item) => item.fieldId === field.id);
-  if (rawValue === null || typeof rawValue === 'undefined' || rawValue === '') {
-    if (index >= 0) current.splice(index, 1);
-    return current;
-  }
-
-  const next: TaskCustomFieldValue = {
-    id: current[index]?.id ?? `temp-${task.id}-${field.id}`,
-    taskId: task.id,
-    fieldId: field.id,
-    field: {
-      id: field.id,
-      name: field.name,
-      type: field.type,
-      required: field.required,
-      position: field.position,
-    },
-    optionId: null,
-    option: null,
-    valueText: null,
-    valueNumber: null,
-    valueDate: null,
-    valueBoolean: null,
-  };
-
-  if (field.type === 'TEXT') {
-    next.valueText = String(rawValue);
-  } else if (field.type === 'NUMBER') {
-    next.valueNumber = Number(rawValue);
-  } else if (field.type === 'DATE') {
-    next.valueDate = String(rawValue);
-  } else if (field.type === 'BOOLEAN') {
-    next.valueBoolean = Boolean(rawValue);
-  } else if (field.type === 'SELECT') {
-    const optionId = String(rawValue);
-    const option = field.options.find((candidate) => candidate.id === optionId) ?? null;
-    next.optionId = optionId;
-    next.option = option
-      ? { id: option.id, label: option.label, value: option.value, color: option.color ?? null }
-      : null;
-    next.valueText = option?.value ?? null;
-  }
-
-  if (index >= 0) current[index] = next;
-  else current.push(next);
-  return current;
-}
 
 type CustomFieldEditorDraft = {
   name: string;
@@ -1322,7 +1168,7 @@ export default function ProjectBoard({
     }),
   );
   const queryClient = useQueryClient();
-  const [selectedTaskId, setSelectedTaskId] = useState<string | null>(null);
+  const { selectedTaskId, setSelectedTaskId, moveTaskPreview } = useProjectBoardState();
   const [collapsedTaskIds, setCollapsedTaskIds] = useState<Set<string>>(new Set());
   const [collapsedSectionIds, setCollapsedSectionIds] = useState<Set<string>>(new Set());
   const [highlightedSectionId, setHighlightedSectionId] = useState<string | null>(null);
