+++ apps/web-ui/src/components/project-board/project-board-state.ts
@@ -1,0 +1,45 @@
+'use client';
+
+import { useState } from 'react';
+import type { SectionTaskGroup, Task } from '@/lib/types';
+
+function moveTaskPreview(
+  groups: SectionTaskGroup[],
+  taskId: string,
+  toSectionId: string,
+  targetTaskId: string | null,
+) {
+  const cloned = groups.map((group) => ({ ...group, tasks: [...group.tasks] }));
+  let movingTask: Task | null = null;
+
+  for (const group of cloned) {
+    const index = group.tasks.findIndex((task) => task.id === taskId);
+    if (index >= 0) {
+      movingTask = group.tasks[index] ?? null;
+      group.tasks.splice(index, 1);
+      break;
+    }
+  }
+
+  if (!movingTask) return groups;
+
+  const destination = cloned.find((group) => group.section.id === toSectionId);
+  if (!destination) return groups;
+
+  const insertAt = targetTaskId
+    ? Math.max(destination.tasks.findIndex((task) => task.id === targetTaskId), 0)
+    : destination.tasks.length;
+
+  destination.tasks.splice(insertAt, 0, { ...movingTask, sectionId: toSectionId });
+  return cloned;
+}
+
+export function useProjectBoardState() {
+  const [selectedTaskId, setSelectedTaskId] = useState<string | null>(null);
+
+  return {
+    selectedTaskId,
+    setSelectedTaskId,
+    moveTaskPreview,
+  };
+}
